app/requirements.txt
# app/requirements.txt
mysql-connector-python

